+++ resources/views/companies/edit.blade.php
@@ -1,0 +1,4 @@
+@extends('layouts.app')
+@section('content')
+    <h1>Editar Empresa</h1>
+@endsection