+++ resources/views/companies/show.blade.php
@@ -1,0 +1,4 @@
+@extends('layouts.app')
+@section('content')
+    <h1>Detalhes da Empresa</h1>
+@endsection