+++ resources/views/customers/show.blade.php
@@ -1,0 +1,4 @@
+@extends('layouts.app')
+@section('content')
+    <h1>Detalhes do Cliente</h1>
+@endsection